Removed ConfigMappedNameResolverFactory
Previously, configuration-based mapping had to recreate parts of the
internal logic of grpc, which deals with default schemes.
It also contained logic for a default address if the address was not
specified. This was moved together with the address selection from the
configuration to the channel factories. Like grpc intended it.

This removes the dependency of the factories on the name resolvers and
allows to resolve them a bit later.
The drawback is that the configuration must be populated before the
factories, but Spring populates it at the earliest possible time anyway.