Purge bearer auth example
The example is mostly non functional and just statically provides the
bearer token, which is usually a dynamic token.

The basic auth example and the docs accurately describe the security
setup. We could still add a bearer auth setup using i.e. keycloak, but
then it should be included in the testExamples.sh script to ensure it
doesn't break unnoticed.